Reorganize tree-related snapshot code
